research: rebuild and test
- rebuilt emHdBindings files with new code
- updated function calls containing FS

NOTE: this commit will not load any materials. there is a bug that shows the following error:

`Warning: in UsdImagingBuildHdMaterialNetworkFromTerminal at line 397 of /Users/andrewbeers/git/projects/emscripten/webgpu/USD/pxr/usdImaging/usdImaging/materialParamUtils.cpp -- Invalid info:id UsdPreviewSurface node:<path/of/material>`

it appears the following line of c++ code fails:

if (!shaderReg.GetNodeByIdentifier(terminalNode.identifier))

here terminalNode is the material node, terminalNode.identifer is `UsdPreviewSurface` and shaderReg is the result of calling SdrRegistry::GetInstance(). Since UsdPreviewSurface is a valid info:id, it appears to me the SdrRegistry has not been initialized properly. I'm not sure what is required to do this as in the current main branch, this seems to work fine and in both branches the only initialization process I see is `getUsdModule`. There are some references to usdImaging in the emHdBindings files however they are the same in this branch and the main branch.